Add number of significant figures parameter
The parameter 'nsf' (number of significant figures) was added to the
functions with the 'bounds' parameter, i.e. bootstrap_f2(), mimcr(),
f1(), f2() and get_profile_portion(). The 'nsf' parameter allows to
control how the test values are rounded before they are compared with
the limits set by the 'bounds' parameter. For example, by setting the
nsf value for the comparison with the upper bound of 85% to 2, an
observed test value of 85.025% is rounded to two significant figures
before the comparison takes place. Thus, the test value is evaluated
as not being greater than the limit.